pyproject.toml
[project]
name = "mem"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioconsole>=0.8.2",
    "dspy>=3.0.3",
    "mem0ai>=0.1.118",
    "openai>=1.109.1",
    "qdrant-client>=1.15.1",
]

[tool.setuptools]
packages = ["mem"]
